+++ minigu/storage/src/tp/mod.rs
@@ -1,12 +1,12 @@
 pub mod checkpoint;
 pub mod iterators;
+pub mod manager;
 pub mod memory_graph;
 pub mod transaction;
-pub mod txn_manager;
 pub mod vector_index;
 
 // Re-export commonly used types for OLTP
+pub use manager::MemTxnManager;
 pub use memory_graph::MemoryGraph;
 pub use transaction::{GraphTxnView, MemTransaction};
-pub use txn_manager::MemTxnManager;
 pub use vector_index::{InMemANNAdapter, VectorIndex};
